FrontEnd
The Hotel Reservation System is a modern web-based platform designed to manage hotel reservations and bookings. The system integrates real-time hotel availability, reservation management, and payment processing, providing users with an intuitive interface for booking hotels. It also offers administrators features to manage hotel details, availability, and customer interactions. Built using Spring Boot for the backend and React for the frontend, the application ensures a seamless user experience and scalability. MySQL is used as the database management system for storing hotel, booking, and customer data.